Move _to_datetime to util.py
