+++ forest/drivers/gridded_forecast.py
@@ -13,6 +13,7 @@ import cftime
 import glob
 from forest import geo
 from forest.view import UMView
+from forest.util import to_datetime as _to_datetime
 
 
 def empty_image():
@@ -29,25 +30,6 @@ def empty_image():
         "length": [],
         "level": []
     }
-
-
-def _to_datetime(d):
-
-    if isinstance(d, datetime):
-        return d
-    if isinstance(d, cftime.DatetimeNoLeap):
-        return datetime(d.year, d.month, d.day, d.hour, d.minute, d.second)
-    elif isinstance(d, cftime.DatetimeGregorian):
-        return datetime(d.year, d.month, d.day, d.hour, d.minute, d.second)
-    elif isinstance(d, str):
-        try:
-            return datetime.strptime(d, "%Y-%m-%d %H:%M:%S")
-        except ValueError:
-            return datetime.strptime(d, "%Y-%m-%dT%H:%M:%S")
-    elif isinstance(d, np.datetime64):
-        return d.astype(datetime)
-    else:
-        raise Exception("Unknown value: {} type: {}".format(d, type(d)))
 
 
 def coordinates(valid_time, initial_time, pressures, pressure):
