+++ forest/util.py
@@ -1,6 +1,7 @@
 import os
 import re
 import datetime as dt
+import cftime
 from functools import partial
 import scipy.ndimage
 import numpy as np
@@ -43,3 +44,24 @@ def initial_time(path):
     groups = re.search(r"[0-9]{8}T[0-9]{4}Z", path)
     if groups:
         return dt.datetime.strptime(groups[0], "%Y%m%dT%H%MZ")
+
+
+def to_datetime(d):
+
+    if isinstance(d, dt.datetime):
+        return d
+    if isinstance(d, cftime.DatetimeNoLeap):
+        return dt.datetime(d.year, d.month, d.day, d.hour, d.minute, d.second)
+    elif isinstance(d, cftime.DatetimeGregorian):
+        return dt.datetime(d.year, d.month, d.day, d.hour, d.minute, d.second)
+    elif isinstance(d, str):
+        try:
+            return dt.datetime.strptime(d, "%Y-%m-%d %H:%M:%S")
+        except ValueError:
+            return dt.datetime.strptime(d, "%Y-%m-%dT%H:%M:%S")
+    elif isinstance(d, np.datetime64):
+        return d.astype(dt.datetime)
+    else:
+        raise Exception("Unknown value: {} type: {}".format(d, type(d)))
+
+
